Incorporated logic for the endpoint All states

+++ FoodOrderingApp-service/src/main/java/com/upgrad/FoodOrderingApp/service/businness/AddressService.java
@@ -96,12 +96,24 @@ public class AddressService {
 
     @Transactional(propagation = Propagation.REQUIRED)
     public String deleteAddress(AddressEntity addressEntity, CustomerEntity signedInCustomer, CustomerEntity belongsToAddressEntity)throws AuthorizationFailedException{
-        if(signedInCustomer.getUuid() != belongsToAddressEntity.getUuid()){
-            throw new AuthorizationFailedException("ATHR-004", "You are not authorized to view/update/delete any one else's address ");
+        if(addressEntity.getActive() == 0){
+            if(signedInCustomer.getUuid() != belongsToAddressEntity.getUuid()){
+                throw new AuthorizationFailedException("ATHR-004", "You are not authorized to view/update/delete any one else's address ");
+            }
+            else{
+                return addressDao.deleteAddress(addressEntity);
+            }
         }
-        else{
-            return addressDao.deleteAddress(addressEntity);
+        else {
+            addressEntity.setActive(0);
+            return null; //need to check on this
         }
+
+    }
+
+    @Transactional(propagation = Propagation.REQUIRED)
+    public List<StateEntity> getAllStates(){
+        return stateDao.getAllStates();
     }
 
 }
